#4331: Added stablehlo builder python infra and cleaned up builder APIs.
