+++ docs/sphinx/builder/builder.rst
@@ -1,0 +1,15 @@
+builder.apis
+==================
+
+.. autoclass:: builder.base.builder.TypeInfo
+   :members:
+
+.. autoclass:: builder.base.builder.Golden
+   :members:
+
+.. autoclass:: builder.base.builder.GoldenCheckLevel
+   :members:
+
+.. autodata:: builder.base.builder.Operand
+
+.. autodata:: builder.base.builder.Shape
