+++ docs/sphinx/builder/stablehlo-builder.rst
@@ -1,0 +1,5 @@
+StableHLOBuilder.apis
+==================
+
+.. autoclass:: builder.stablehlo.stablehlo_builder.StableHLOBuilder
+   :members:
