+++ docs/sphinx/builder/stablehlo-utils.rst
@@ -1,0 +1,4 @@
+builder.stablehlo.stablehlo\_utils
+===================
+
+.. autofunction:: builder.stablehlo.stablehlo_utils.build_stablehlo_module
