+++ docs/sphinx/builder/ttir-builder.rst
@@ -1,0 +1,5 @@
+TTIRBuilder.apis
+==================
+
+.. autoclass:: builder.ttir.ttir_builder.TTIRBuilder
+   :members:
