+++ docs/sphinx/builder/ttir-golden.rst
@@ -1,3 +1,6 @@
+ttir\_golden
+===========================
+
 ttir\_builder.ttir\_golden
 ===========================
