+++ docs/sphinx/builder/ttir-utils.rst
@@ -1,0 +1,10 @@
+builder.ttir.ttir\_utils
+===================
+
+.. autofunction:: builder.ttir.ttir_utils.build_ttir_module
+
+.. autofunction:: builder.ttir.ttir_utils.create_custom_ttir_pipeline_fn
+
+.. autofunction:: builder.ttir.ttir_utils.run_ttir_pipeline
+
+.. autofunction:: builder.ttir.ttir_utils.compile_ttir_to_flatbuffer
