--- docs/sphinx/ttir-builder/apis.rst
@@ -1,14 +1,0 @@
-ttir\_builder.apis
-==================
-
-.. autoclass:: ttir_builder.Golden
-   :members:
-
-.. autoclass:: ttir_builder.GoldenCheckLevel
-   :members:
-
-.. autoclass:: ttir_builder.TypeInfo
-   :members:
-
-.. autoclass:: ttir_builder.TTIRBuilder
-    :members:
