--- docs/sphinx/ttir-builder/ops.rst
@@ -1,5 +1,0 @@
-ttir\_builder.ops
-=================
-
-.. autoclass:: ttir_builder.TTIRBuilderOps
-   :members:
