--- docs/sphinx/ttir-builder/utils.rst
@@ -1,16 +1,0 @@
-ttir\_builder.utils
-===================
-
-.. autofunction:: ttir_builder.utils.shape_str
-
-.. autofunction:: ttir_builder.utils.set_output_path
-
-.. autofunction:: ttir_builder.utils.get_target_path
-
-.. autofunction:: ttir_builder.utils.create_custom_pipeline_fn
-
-.. autofunction:: ttir_builder.utils.build_mlir_module
-
-.. autofunction:: ttir_builder.utils.run_pipeline
-
-.. autofunction:: ttir_builder.utils.compile_to_flatbuffer
