+++ docs/src/SUMMARY.md
@@ -14,8 +14,9 @@
   - [`ttmlir-opt`](./ttmlir-opt.md)
   - [`ttmlir-translate`](./ttmlir-translate.md)
   - [`ttrt`](./ttrt.md)
-  - [`ttir-builder`](./ttir-builder/ttir-builder.md)
-     - [Adding a TTIR op](./ttir-builder/adding-a-ttir-op.md)
+  - [`ttir-builder`](./builder/ttir-builder.md)
+     - [Adding a TTIR op](./builder/adding-a-ttir-op.md)
+  - [`stablehlo-builder`](./builder/stablehlo-builder.md)
   - [`tt-explorer`](./tt-explorer/tt-explorer.md)
     - [UI & changes to Model Explorer](./tt-explorer/ui.md)
     - [CLI](./tt-explorer/cli.md)
