+++ test/python/golden/test_metal_tilize.py
@@ -6,10 +6,15 @@ import pytest
 import torch
 from typing import Callable, List
 
-from ttir_builder.utils import compile_to_flatbuffer, Marks, shape_str
-from ttir_builder import Operand, TTIRBuilder, UnitAttr, Shape, TypeInfo, ttir_golden
 from ttmlir.dialects import ttir, ttcore
 from ttmlir.ir import *
+
+from builder.base.builder import Operand, Shape
+from builder.ttir.ttir_builder import TTIRBuilder
+from builder.ttir import ttir_golden
+from builder.ttir.ttir_utils import compile_ttir_to_flatbuffer
+
+from test_utils import Marks, shape_str
 
 
 @pytest.mark.parametrize("shape", [(32, 64), (64, 32), (64, 64), (64, 128)])
@@ -22,13 +27,13 @@ def test_tilize(shape: Shape, request):
 
         to_device = builder.tilize(
             in0,
-            output_type=builder.metal_tensor_layout(shape, tiled=True),
+            output_type=builder.get_metal_tensor_layout(shape, tiled=True),
             unit_attrs=unit_attrs,
         )
 
         view_as_rm = builder.view_layout(
             to_device,
-            output_type=builder.metal_tensor_layout(shape, tiled=False),
+            output_type=builder.get_metal_tensor_layout(shape, tiled=False),
             reinterpret_layout=True,
             unit_attrs=unit_attrs,
         )
@@ -41,7 +46,7 @@ def test_tilize(shape: Shape, request):
 
         return from_device
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         tilize,
         [shape],
         target="ttmetal",
@@ -71,13 +76,13 @@ def test_untilize(shape: Shape, request):
 
         to_device = builder.to_layout(
             in0,
-            output_type=builder.metal_tensor_layout(shape, (1, 1), False),
+            output_type=builder.get_metal_tensor_layout(shape, (1, 1), False),
             unit_attrs=unit_attrs,
         )
 
         view_as_tiled = builder.view_layout(
             to_device,
-            output_type=builder.metal_tensor_layout(shape, (1, 1), True),
+            output_type=builder.get_metal_tensor_layout(shape, (1, 1), True),
             reinterpret_layout=True,
             unit_attrs=unit_attrs,
         )
@@ -90,7 +95,7 @@ def test_untilize(shape: Shape, request):
 
         return from_device
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         untilize,
         [shape],
         target="ttmetal",
@@ -110,7 +115,7 @@ def test_tilize_untilize(shape: Shape, request):
     ):
         to_device = builder.tilize(
             in0,
-            output_type=builder.metal_tensor_layout(shape, (1, 1), True),
+            output_type=builder.get_metal_tensor_layout(shape, (1, 1), True),
             unit_attrs=unit_attrs,
         )
         from_device = builder.untilize(
@@ -120,7 +125,7 @@ def test_tilize_untilize(shape: Shape, request):
         )
         return from_device
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         tilize_untilize,
         [shape],
         target="ttmetal",
