+++ test/python/golden/test_ttir_ops_llmbox.py
@@ -5,8 +5,10 @@ import torch
 import pytest
 
 from typing import List, Tuple
-from ttir_builder.utils import compile_to_flatbuffer
-from ttir_builder import Operand, TTIRBuilder, Shape
+
+from builder.base.builder import Operand, Shape
+from builder.ttir.ttir_builder import TTIRBuilder
+from builder.ttir.ttir_utils import compile_ttir_to_flatbuffer
 
 pytestmark = pytest.mark.llmbox
 
@@ -66,7 +68,7 @@ def test_all_gather(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(2, -1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         all_gather, [shape], mesh_shape=mesh_shape, test_base=request.node.name
     )
 
@@ -119,7 +121,7 @@ def test_all_reduce(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(2, -1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         all_reduce, [shape], mesh_shape=mesh_shape, test_base=request.node.name
     )
 
@@ -186,7 +188,7 @@ def test_reduce_scatter(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(2, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         reduce_scatter,
         [shape],
         mesh_shape=mesh_shape,
@@ -260,7 +262,7 @@ def test_collective_permute(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(2, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         collective_permute,
         [shape],
         mesh_shape=mesh_shape,
@@ -335,7 +337,7 @@ def test_matmul_2x4(shapes: List[Shape], mesh_shape: Tuple[int, int], request):
             shard_dims=(0, -1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         matmul_2x4,
         shapes,
         mesh_shape=mesh_shape,
@@ -405,7 +407,7 @@ def test_matmul_1x8(shapes: List[Shape], mesh_shape: Tuple[int, int], request):
             shard_dims=(-1,),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         matmul_1x8,
         shapes,
         mesh_shape=mesh_shape,
@@ -453,7 +455,7 @@ def test_neg_2x4(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(1, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_2x4,
         [shape],
         mesh_shape=mesh_shape,
@@ -501,7 +503,7 @@ def test_neg_2x4_cluster_0(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(1, -1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_2x4_cluster_0,
         [shape],
         mesh_shape=mesh_shape,
@@ -549,7 +551,7 @@ def test_neg_2x4_cluster_1(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(1, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_2x4_cluster_1,
         [shape],
         mesh_shape=mesh_shape,
@@ -597,7 +599,7 @@ def test_neg_2x4_reversed_cluster(shape: Shape, mesh_shape: Tuple[int, int], req
             shard_dims=(3, 1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_2x4_reversed_cluster,
         [shape],
         mesh_shape=mesh_shape,
@@ -645,7 +647,7 @@ def test_neg_2x4_reversed_cluster_0(shape: Shape, mesh_shape: Tuple[int, int], r
             shard_dims=(3, -1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_2x4_reversed_cluster_0,
         [shape],
         mesh_shape=mesh_shape,
@@ -694,7 +696,7 @@ def test_neg_1x8_dim_3(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(-1, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_1x8_dim_3,
         [shape],
         mesh_shape=mesh_shape,
@@ -743,7 +745,7 @@ def test_neg_1x8_dim_1(shape: Shape, mesh_shape: Tuple[int, int], request):
             shard_dims=(-1, 1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         neg_1x8_dim_1,
         [shape],
         mesh_shape=mesh_shape,
@@ -800,7 +802,7 @@ def test_eltwise_multidevice(shapes: List[Shape], mesh_shape: Tuple[int, int], r
             shard_dims=(0, 1),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         eltwise_multidevice,
         shapes,
         mesh_shape=mesh_shape,
@@ -856,7 +858,7 @@ def test_matmul_and_binary_op(
         output = builder.add(unsharded, in2)
         return output
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         matmul_test,
         shapes,
         mesh_shape=mesh_shape,
@@ -911,7 +913,7 @@ def test_matmul_and_unary_op(shapes: List[Shape], mesh_shape: Tuple[int, int], r
         output = builder.neg(unsharded)
         return output
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         matmul_test,
         shapes,
         mesh_shape=mesh_shape,
@@ -1004,7 +1006,7 @@ def test_matmul_and_binary_op_2(
         output = builder.add(matmul_0, matmul_2)
         return output
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         matmul_test,
         shapes,
         mesh_shape=mesh_shape,
@@ -1098,7 +1100,7 @@ def all_to_all_test(
             shard_dims=shard_dims,
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         all_to_all,
         [input_shape],
         mesh_shape=mesh_shape,
@@ -1265,7 +1267,7 @@ def test_collective_broadcast(
             shard_dims=(2, 3),
         )
 
-    compile_to_flatbuffer(
+    compile_ttir_to_flatbuffer(
         collective_broadcast,
         [shape],
         mesh_shape=mesh_shape,
