+++ tools/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_subdirectory(ttmlir-opt)
 add_subdirectory(ttmlir-lsp-server)
 if(TTMLIR_ENABLE_BINDINGS_PYTHON AND MLIR_ENABLE_BINDINGS_PYTHON)
-  add_subdirectory(ttir-builder)
+  add_subdirectory(builder)
 
   if (TTMLIR_ENABLE_PYKERNEL)
     add_subdirectory(pykernel)
