+++ tools/builder/CMakeLists.txt
@@ -1,0 +1,26 @@
+include(AddMLIRPython)
+
+declare_mlir_python_sources(BuilderSources
+  ROOT_DIR "${CMAKE_CURRENT_SOURCE_DIR}"
+  SOURCES
+    # builder
+    base/__init__.py
+    base/builder.py
+
+    # ttir subpackage files
+    ttir/__init__.py
+    ttir/ttir_builder.py
+    ttir/ttir_utils.py
+    ttir/ttir_golden.py
+
+    # stablehlo subpackage files
+    stablehlo/__init__.py
+    stablehlo/stablehlo_builder.py
+    stablehlo/stablehlo_utils.py
+)
+
+add_mlir_python_modules(BuilderPythonModules
+  ROOT_PREFIX "${TTMLIR_PYTHON_PACKAGES_DIR}/builder"
+  INSTALL_PREFIX "python_packages/builder"
+  DECLARED_SOURCES BuilderSources
+)
