+++ tools/builder/README.md
@@ -1,0 +1,5 @@
+# `builder`
+
+`builder` contains `ttir-builder` and `stablehlo-builder` to create ttir or stablehlo graphs.
+
+To get started, refer to `builder` [documentation](../../docs/src/builder/).
