+++ tools/builder/__init__.py
@@ -1,0 +1,21 @@
+# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+from .base.builder import Builder, Operand, Shape, TypeInfo, Golden, GoldenCheckLevel
+from .ttir.ttir_builder import TTIRBuilder
+from .stablehlo.stablehlo_builder import StableHLOBuilder
+from .ttir import ttir_utils
+from .stablehlo import stablehlo_utils
+
+__all__ = [
+    "Builder",
+    "Operand",
+    "Shape",
+    "TypeInfo",
+    "Golden",
+    "GoldenCheckLevel",
+    "TTIRBuilder",
+    "StableHLOBuilder",
+    "ttir_utils",
+    "stablehlo_utils",
+]
