+++ tools/builder/ttir/ttir_golden.py
