--- tools/ttir-builder/__init__.py
@@ -1,11 +1,0 @@
-# SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-"""
-TTIR Builder
-
-This module contains the `TTIRBuilder` class, which is the main entry point for creating TTIR from python.
-
-"""
-from .apis import *
-from .ops import *
