--- tools/ttir-builder/release_notes.md
@@ -1,10 +1,0 @@
-# `ttir-builder`
-
-v1.0.0
-June 27th, 2025
-
-### New features
- - [Building a TTIR module](../../docs/src/ttir-builder/ttir-builder.md#creating-a-ttir-module)
- - [Running pipelines](../../docs/src/ttir-builder/ttir-builder.md#running-a-pipeline)
- - [Compiling modules to flatbuffers](../../docs/src/ttir-builder/ttir-builder.md#compiling-into-flatbuffer)
- - [Golden accuracy checks](../../docs/src/ttir-builder/ttir-builder.md#golden-mode)
